* release/201208030130: (278 commits)
  Centering the Sign Up button on the Essential Guide landing page
  Completing the work for Essential Guide social login changes (to-do: https://basecamp.com/1759305/projects/19290-collectorsquest-com/todos/11686264-make-the-social)
  Reverting the latest changes to _footer_login_signup.php and _footer_signup_external_buttons.php
  Tabs -> Spaces
  Rename "Uncategorized" ContentCategory to bottom as "Other"
  Prefix custom required|invalid errors with the field name
  Update tags required error copy for Collection/Collectible
  Show all errors for collection at the top of the page
  Better formatting for default symfony errors (Invalid|Required) in Bootstrap formatter
  Implemented conditional opening and closing of the Dropbox in MyCQ
  Adding @method comments for sfAction::redirect() to avoid IDE code inspection warnings
  Switching the classes for "Save Changes" and "Save & Go to Items" buttons
  Renamed the cookie "cqRemember" to "cq_remember"
  fix to-do: https://basecamp.com/1759305/projects/19290-collectorsquest-com/todos/10806532-fixed-feedback label wrapper and send new email instance instead Cc
  If form thumbnail is set it will always be an instance of sfValidatedFile, no need to check that
  Fix E_NOTICE in collectionAction.class.php L107
  Fixing a problem with the shadow of the MyCQ tabs (example: http://www.evernote.com/shard/s154/sh/b90e14c7-c12c-40de-ade1-203085873517/647de99979ff5acb1a3fc4556e2e5ae3)
  fix to-do: https://basecamp.com/1759305/projects/19290-collectorsquest-com/todos/10806532-fixed-feedback fix bug with sending email when send copy is off
  Added "Follow us on Google+ " to the footer
  Set up the tooltips to show below icons to-do https://basecamp.com/1759305/projects/19290-collectorsquest-com/todos/11867779-we-need-to-add
  ...

Conflicts:
	apps/frontend/lib/cqFrontendUser.class.php
	apps/frontend/lib/form/CollectorGuideSignupForm.class.php
	apps/frontend/lib/form/CollectorSignupStep1Form.class.php
	apps/frontend/modules/misc/templates/_guide_footer.php

+++ apps/frontend/modules/mycq/actions/ajaxAction.class.php
@@ -183,6 +183,9 @@ class ajaxAction extends cqAjaxAction
   }
 
   /**
+   * @param  sfWebRequest  $request
+   *
+   * @throws PropelException
    * @return string
    */
   protected function executeCollectibleDonateImage(sfWebRequest $request)
@@ -215,11 +218,16 @@ class ajaxAction extends cqAjaxAction
       {
         $primary->delete();
       }
+      else if (!$is_primary && !$recipient->getPrimaryImage())
+      {
+        $is_primary = true;
+      }
 
       try
       {
         $image->setIsPrimary($is_primary);
         $image->setModelId($recipient->getId());
+        $image->setSource($donor->getId());
         $image->save();
       }
       catch (PropelException $e)
@@ -237,7 +245,16 @@ class ajaxAction extends cqAjaxAction
       $recipient->setUpdatedAt(time());
       $recipient->save();
 
-      // Delete the $donor, not needed anymore
+      // auto-set collection thumbnail if none set yet
+      $collection = $recipient->getCollectorCollection();
+      if (1 == $collection->countCollectibles() && !$collection->hasThumbnail())
+      {
+        $collection->setPrimaryImage($recipient->getPrimaryImage()
+          ->getAbsolutePath('original'));
+        $collection->save();
+      }
+
+      // Archive the $donor, not needed anymore
       $donor->delete();
 
       // Return "Success"
@@ -289,6 +306,7 @@ class ajaxAction extends cqAjaxAction
   }
 
   /**
+   * @param  sfWebRequest  $request
    * @return string
    */
   protected function executeCollectionSetThumbnail(sfWebRequest $request)
@@ -321,6 +339,8 @@ class ajaxAction extends cqAjaxAction
 
     if ($multimedia instanceof iceModelMultimedia)
     {
+      $object = $multimedia->getModelObject();
+
       if ($multimedia->getIsPrimary())
       {
         $model = $multimedia->getModelObject();
@@ -331,7 +351,35 @@ class ajaxAction extends cqAjaxAction
         }
       }
 
-      $multimedia->delete();
+      /** @var $archive CollectibleArchive */
+      if (
+        ($source = $multimedia->getSource()) &&
+        ($archive = CollectibleArchiveQuery::create()->findOneById((integer) $source))
+      )
+      {
+        $collectible = new Collectible();
+        $collectible->populateFromArchive($archive);
+        $collectible->save();
+
+        $multimedia->setModel($collectible);
+        $multimedia->setSource(null);
+        $multimedia->save();
+
+        /**
+         * Update the Eblob cache
+         */
+        if ($object)
+        {
+          $m = iceModelMultimediaPeer::retrieveByModel($object);
+
+          $object->setEblobElement('multimedia', $m->toXML(true));
+          $object->save();
+        }
+      }
+      else
+      {
+        $multimedia->delete();
+      }
     }
 
     return $this->success();
@@ -383,4 +431,185 @@ class ajaxAction extends cqAjaxAction
 
     return $this->success();
   }
+
+  /**
+   * section: collection
+   * page: createStep1
+   */
+  protected function executeCollectionCreateStep1(sfWebRequest $request, $template)
+  {
+    $form = new CollectionCreateForm();
+    $form->setDefault('collectible_id', $request->getParameter('collectible_id'));
+
+    if (sfRequest::POST == $request->getMethod())
+    {
+      $form->bind($request->getParameter('collection'));
+      if ($form->isValid())
+      {
+        $values = $form->getValues();
+        $values['collector_id'] = $this->getUser()->getCollector()->getId();
+
+        /** @var $collection CollectorCollection */
+        $collection = $form->updateObject($values);
+        $collection->setTags($values['tags']);
+        $collection->save();
+
+        if (isset($values['collectible_id']))
+        {
+          $q = CollectibleQuery::create()
+            ->filterByCollector($this->getUser()->getCollector())
+            ->filterById($values['collectible_id']);
+
+          if (($collectible = $q->findOne()) && $this->getUser()->isOwnerOf($collectible))
+          {
+            // Let's create the CollectionCollectible
+            $q = CollectionCollectibleQuery::create()
+              ->filterByCollection($collection)
+              ->filterByCollectible($collectible);
+
+            $collection_collectible = $q->findOneOrCreate();
+            $collection_collectible->save();
+
+            /**
+             * If the Collectible has a thumnail (it should!),
+             * let's add it as the Collection thumbnail also
+             *
+             * @var $thumbnail iceModelMultimedia
+             */
+            if ($thumbnail = $collectible->getPrimaryImage())
+            {
+              $collection->setThumbnail($thumbnail->getAbsolutePath('original'));
+              $collection->save();
+            }
+
+          }
+        }
+
+        $this->getUser()->getCollector()->getProfile()->updateProfileProgress();
+
+        return $this->redirect('ajax_mycq', array(
+            'section' => 'collection',
+            'page' => 'createStep2',
+            'collection_id' => $collection->getId(),
+        ));
+      }
+    }
+
+    $root = ContentCategoryQuery::create()->findRoot();
+    $this->categories = ContentCategoryQuery::create()
+        ->descendantsOf($root)
+        ->findTree();
+
+    $this->form = $form;
+
+    return $template;
+  }
+
+  /**
+   * section: collection
+   * page: createStep2
+   * params: collection_id
+   */
+  public function executeCollectionCreateStep2(sfWebRequest $request, $template)
+  {
+    $collection = CollectorCollectionPeer::retrieveByPK(
+      $request->getParameter('collection_id')
+    );
+    $this->forward404Unless($collection &&
+      $this->getUser()->getCollector()->isOwnerOf($collection));
+
+    $form = new CollectorCollectionEditForm($collection);
+    if ($collection->hasThumbnail())
+    {
+      $form->useFields(array(
+          'description',
+      ));
+    }
+    else
+    {
+      $form->useFields(array(
+          'thumbnail',
+          'description',
+      ));
+    }
+
+    if (sfRequest::POST == $request->getMethod())
+    {
+      $taintedValues = $request->getParameter($form->getName());
+      $form->bind($taintedValues, $request->getFiles($form->getName()));
+
+      if ($form->isValid())
+      {
+        $values = $form->getValues();
+        $collection->setDescription($values['description'], 'html');
+
+        if (isset($values['thumbnail']))
+        {
+          $collection->setThumbnail($values['thumbnail']);
+        }
+
+        $collection->save();
+
+        // Tell the Dropbox to stay closed
+        $this->getUser()->setFlash('cq_mycq_dropbox_open', false, true, 'cookies');
+
+        return $this->renderPartial('global/loading', array(
+            'url' => $this->generateUrl('mycq_collection_by_section', array(
+                'id' => $collection->getId(),
+                'section' => 'details',
+            )),
+        ));
+      }
+    }
+
+    $this->form = $form;
+    $this->collection = $collection;
+
+    return $template;
+  }
+
+  /**
+   * section: collection
+   * page: changeCategoty
+   * params: collection_id
+   */
+  public function executeCollectionChangeCategory(sfWebRequest $request, $template)
+  {
+    $collection = CollectorCollectionPeer::retrieveByPK(
+      $request->getParameter('collection_id')
+    );
+    $this->forward404Unless($this->getUser()->isOwnerOf($collection));
+
+    $form = new CollectionCreateForm($collection);
+    $form->useFields(array('content_category_id'));
+
+    if (sfRequest::POST == $request->getMethod())
+    {
+      $taintedValues = $request->getParameter($form->getName());
+      $form->bind($taintedValues, $request->getFiles($form->getName()));
+
+      if ($form->isValid())
+      {
+        $form->save();
+
+        return $this->renderPartial('global/loading', array(
+            'url' => $this->generateUrl('mycq_collection_by_section', array(
+                'id' => $collection->getId(),
+                'section' => 'details',
+            )),
+        ));
+      }
+    }
+
+    $this->form = $form;
+    $this->collection = $collection;
+
+    $root = ContentCategoryQuery::create()->findRoot();
+    $this->categories = ContentCategoryQuery::create()
+        ->descendantsOf($root)
+        ->findTree();
+
+    return $template;
+  }
+
 }
